Fix -Wold-style-cast
Additionally, add a nocompile test for flags limited to specific
file in the warning_suppression_mappings.txt that should still
trigger warnings for other code, using -Wold-style-cast.

Add warning_suppression_mappings.txt as an input so that when it
is modified a rebuild is triggered.

Bug: 523358231
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/323178
Commit-Queue: Kai Ninomiya <kainino@chromium.org>

+++ src/dawn/tests/end2end/BufferHostMappedPointerTests_posix.cpp
@@ -62,7 +62,7 @@ class MMapBackend : public BufferHostMappedPointerTestBackend {
         // Write the initial data.
         std::vector<char> initialData(size);
         Populate(initialData.data());
-        EXPECT_EQ(write(fd, initialData.data(), size), (signed)size);
+        EXPECT_EQ(write(fd, initialData.data(), size), static_cast<ssize_t>(size));
 
         // Memory map the file.
         void* ptr = mmap(nullptr, size, PROT_READ | PROT_WRITE, MAP_PRIVATE, fd, 0);
